+++ src/Tests/ServiceBus.fs
@@ -12,6 +12,19 @@ open Microsoft.Azure.Management.ServiceBus.Models
 open Microsoft.Rest
 open System
 
+
+let getResource<'T when 'T :> IArmResource> (data:IArmResource list) = data |> List.choose(function :? 'T as x -> Some x | _ -> None)
+let getTopicResource = getResource<Topic>
+
+let getResources (v:IBuilder) = v.BuildResources Location.WestUS
+
+let getResourceDependsOnByName (template:Deployment) (resourceName:ResourceName) =
+    let json = template.Template |> Writer.toJson
+    let jobj = Newtonsoft.Json.Linq.JObject.Parse(json)
+    let dependsOn = jobj.SelectToken($"resources[?(@.name=='{resourceName.Value}')].dependsOn")
+    let jarray = dependsOn :?> Newtonsoft.Json.Linq.JArray
+    [for jvalue in jarray do jvalue.ToString()]
+
 /// Client instance needed to get the serializer settings.
 let dummyClient = new ServiceBusManagementClient (Uri "http://management.azure.com", TokenCredentials "NotNullOrWhiteSpace")
 let getResourceAtIndex o = o |> getResourceAtIndex dummyClient.SerializationSettings
@@ -189,6 +202,27 @@ let tests = testList "Service Bus Tests" [
                 } |> getResourceAtIndex 2
             Expect.equal sub.Name "my-bus/my-topic/my-sub" "Name not set"
         }
+        test "Can create a forwarding subscription" {
+            let sub:SBSubscription =
+                serviceBus {
+                    name "my-bus"
+                    add_topics [
+                        topic {
+                            name "my-topic"
+                            add_subscriptions [
+                                subscription {
+                                    name "my-sub"
+                                    forward_to "my-other-topic"
+                                }
+                            ]
+                        }
+                        topic {
+                            name "my-other-topic"
+                        }
+                    ]
+                } |> getResourceAtIndex 3
+            Expect.equal sub.ForwardTo "my-other-topic" "ForwardTo not set"
+        }
         test "Creates a correlation filter rule" {
             let correlationRule =
                 ServiceBus.CorrelationFilter(
@@ -257,6 +291,111 @@ let tests = testList "Service Bus Tests" [
 
             Expect.hasLength subscriptions 2 "Subscription length"
             Expect.hasLength subscriptions 2 "Subscription length"
+        }
+        test "Topic does not create dependencies for unmanaged linked resources" {
+            let resource =
+                topic {
+                    name "my-topic"
+                    link_to_unmanaged_namespace "my-bus"
+                }
+                |> getResources |> getTopicResource |> List.head
+            Expect.isEmpty resource.Dependencies ""
+        }
+        test "Topic creates dependencies for managed linked resources" {
+            let resource =
+                serviceBus {
+                    name "my-bus"
+                    add_topics [
+                        topic {
+                            name "my-topic"
+                            link_to_unmanaged_namespace "my-namespace"
+                        }        
+                    ]
+                }
+                |> getResources |> getTopicResource |> List.head
+            Expect.containsAll resource.Dependencies [
+                ResourceId.create(namespaces, ResourceName "my-bus");]
+                ""
+        }
+        test "Topic creates empty dependsOn in arm template json for unmanaged linked resources" {
+            let template =
+                arm {
+                    add_resources [
+                        topic {
+                            name "my-topic"
+                            link_to_unmanaged_namespace "my-bus"
+                        }
+                    ]
+                }
+            let dependsOn = getResourceDependsOnByName template (ResourceName "my-bus/my-topic")    
+            Expect.hasLength dependsOn 0 ""
+        }
+        test "Topic creates dependsOn in arm template json for managed linked resources" {
+            let template =
+                arm {
+                    add_resources [
+                        serviceBus {
+                            name "my-bus"
+                            add_topics [
+                                topic {
+                                    name "my-topic"
+                                    link_to_unmanaged_namespace "my-namespace"
+                                }        
+                            ]
+                        }
+                    ]
+                }
+            let dependsOn = getResourceDependsOnByName template (ResourceName "my-bus/my-topic")
+            Expect.hasLength dependsOn 1 ""
+            let expectedNamespaceDependency = "[resourceId('Microsoft.ServiceBus/namespaces', 'my-bus')]"
+            Expect.equal dependsOn.Head expectedNamespaceDependency "" 
+        }
+        test "Topic IBuilder has correct resourceId for unmanaged namespace" {
+            let resource = 
+                topic {
+                    name "my-topic"
+                    link_to_unmanaged_namespace "my-bus"
+                } :> IBuilder
+            Expect.equal (resource.ResourceId.Eval()) "[resourceId('Microsoft.ServiceBus/namespaces/topics', 'my-bus', 'my-topic')]" ""
+        }
+        test "Topic IArmResource has correct resourceId for unmanaged namespace" {
+            let resource = 
+                topic {
+                    name "my-topic"
+                    link_to_unmanaged_namespace "my-bus"
+                }
+                |> getResources |> getTopicResource |> List.head :> IArmResource
+            Expect.equal (resource.ResourceId.Eval()) "[resourceId('Microsoft.ServiceBus/namespaces/topics', 'my-bus', 'my-topic')]" ""
+        }
+        test "Topic IBuilder has correct resourceId for managed namespace" {
+            let topicName = "my-topic"
+            let svcBus =
+                serviceBus {
+                    name "my-bus"
+                    add_topics [
+                        topic {
+                            name topicName
+                            link_to_unmanaged_namespace "other-namespace"
+                        }        
+                    ]
+                }
+            let topicBuilder = svcBus.Topics |> Map.find (ResourceName topicName) :> IBuilder
+            Expect.equal (topicBuilder.ResourceId.Eval()) $"[resourceId('Microsoft.ServiceBus/namespaces/topics', 'my-bus', '{topicName}')]" ""
+        }
+        test "Topic IArmResource has correct resourceId for managed namespace" {
+            let topicName = "my-topic"
+            let resource =
+                serviceBus {
+                    name "my-bus"
+                    add_topics [
+                        topic {
+                            name topicName
+                            link_to_unmanaged_namespace "other-namespace"
+                        }        
+                    ]
+                }
+                |> getResources |> getTopicResource |> List.head :> IArmResource
+            Expect.equal (resource.ResourceId.Eval()) $"[resourceId('Microsoft.ServiceBus/namespaces/topics', 'my-bus', '{topicName}')]" ""
         }
     ]
 ]