+++ RELEASE_NOTES.md
@@ -1,15 +1,32 @@
 Release Notes
 =============
-##  1.5.0-beta
+
+## 1.5.2
+* ServiceBus: TopicConfig implements IBuilder and supports link_to_unmanaged_namespace.
+* ServiceBus: Support for forwarding messages delivered to a subscription.
+
+## 1.5.1
+* Communication Services: add builder.
+* ExpressRoute: Adds ServiceKey property to generate an expression for the service key on a new circuit.
+* Network Security Groups: Enable builder to create outbound rules.
+* ServiceBus: Fix an issue with Premium Sku ARM Writer
+* ServiceBus: Fix an issue with Rules depends on ARM Writer
+* Storage Accounts: Support for CORS.
+* Virtual WAN: add builder
+
+##  1.5.0
 * Container Groups: Support for init containers.
 * Container Groups: Support for liveliness and readiness probes on containers.
 * Container Groups: Connect network profile to an existing virtual network.
 * Container Groups: Bugfix for outputs.
+* CosmosDb: Add support for MongoDB as a database kind.
 * Event Grid: Ensure destination Queues are created as a dependency
 * Event Grid: Add ServiceBus Queue and Topic as supported destinations
 * Functions: Support for 64 bits.
 * Functions: Add option to use managed Key Vault
+* Functions: Add support for dotnet-isolated runtime (NET5)
 * KeyVault: Fix an issue with adding tags on main KeyVault builder.
+* KeyVault: Support Azure RBAC for data plane access.
 * ServiceBus: update namespace validation rules to follow [Microsoft documentation](https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftservicebus)
 * Storage: Add support for tables
 * Web App: Disables the automatic addition of the logging site extension when `docker_image` is used